pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-scraper-pro"
version = "1.0.0"
description = "A responsible, professional web scraper with Excel/CSV output, pagination, and watch mode"
authors = [{ name = "Adrian Azahid García / Strivark" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "openpyxl>=3.1",
    "click>=8.1",
    "tenacity>=8.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
playwright = ["playwright>=1.44"]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-beautifulsoup4>=4.12",
]

[project.scripts]
web-scraper = "web_scraper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/web_scraper"]

[tool.hatch.build.targets.sdist]
include = ["src/", "tests/", "README.md", "LICENSE"]

[tool.ruff]
src = ["src"]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
